Don't delete request on exit if both accepted --> this fixes that only one of two go to match state. The requests need to stay open until the meeting starts

+++ frontend/naklario/src/app/roulette/wait/wait.component.ts
@@ -68,7 +68,9 @@ export class WaitComponent implements OnInit, OnDestroy {
       );
   }
   ngOnDestroy(): void {
-    this.rouletteService.deleteMatch(this.requestType);
+    if (!this.match?.bothAccepted) {
+      this.rouletteService.deleteMatch(this.requestType);
+    }
     this.sub$.unsubscribe();
   }
